+++ core/src/main/java/de/fraunhofer/iosb/ilt/faaast/service/request/handler/aas/PostSubmodelReferenceRequestHandler.java
@@ -18,6 +18,7 @@ import de.fraunhofer.iosb.ilt.faaast.service.exception.MessageBusException;
 import de.fraunhofer.iosb.ilt.faaast.service.model.api.modifier.QueryModifier;
 import de.fraunhofer.iosb.ilt.faaast.service.model.api.request.aas.PostSubmodelReferenceRequest;
 import de.fraunhofer.iosb.ilt.faaast.service.model.api.response.aas.PostSubmodelReferenceResponse;
+import de.fraunhofer.iosb.ilt.faaast.service.model.exception.PersistenceException;
 import de.fraunhofer.iosb.ilt.faaast.service.model.exception.ResourceAlreadyExistsException;
 import de.fraunhofer.iosb.ilt.faaast.service.model.exception.ResourceNotFoundException;
 import de.fraunhofer.iosb.ilt.faaast.service.model.messagebus.event.change.ElementUpdateEventMessage;
@@ -40,7 +41,8 @@ public class PostSubmodelReferenceRequestHandler extends AbstractRequestHandler<
 
 
     @Override
-    public PostSubmodelReferenceResponse process(PostSubmodelReferenceRequest request) throws ResourceNotFoundException, MessageBusException, ResourceAlreadyExistsException {
+    public PostSubmodelReferenceResponse process(PostSubmodelReferenceRequest request)
+            throws ResourceNotFoundException, MessageBusException, ResourceAlreadyExistsException, PersistenceException {
         AssetAdministrationShell aas = context.getPersistence().getAssetAdministrationShell(request.getId(), QueryModifier.DEFAULT);
         if (aas.getSubmodels().contains(request.getSubmodelRef())) {
             throw new ResourceAlreadyExistsException(request.getSubmodelRef());