+++ core/src/main/java/de/fraunhofer/iosb/ilt/faaast/service/request/handler/aasrepository/GetAssetAdministrationShellByIdReferenceRequestHandler.java
@@ -18,6 +18,7 @@ import de.fraunhofer.iosb.ilt.faaast.service.exception.MessageBusException;
 import de.fraunhofer.iosb.ilt.faaast.service.model.api.modifier.QueryModifier;
 import de.fraunhofer.iosb.ilt.faaast.service.model.api.request.aasrepository.GetAssetAdministrationShellByIdReferenceRequest;
 import de.fraunhofer.iosb.ilt.faaast.service.model.api.response.aasrepository.GetAssetAdministrationShellByIdReferenceResponse;
+import de.fraunhofer.iosb.ilt.faaast.service.model.exception.PersistenceException;
 import de.fraunhofer.iosb.ilt.faaast.service.model.exception.ResourceNotFoundException;
 import de.fraunhofer.iosb.ilt.faaast.service.model.messagebus.event.access.ElementReadEventMessage;
 import de.fraunhofer.iosb.ilt.faaast.service.request.handler.AbstractRequestHandler;
@@ -43,7 +44,8 @@ public class GetAssetAdministrationShellByIdReferenceRequestHandler
 
 
     @Override
-    public GetAssetAdministrationShellByIdReferenceResponse process(GetAssetAdministrationShellByIdReferenceRequest request) throws ResourceNotFoundException, MessageBusException {
+    public GetAssetAdministrationShellByIdReferenceResponse process(GetAssetAdministrationShellByIdReferenceRequest request)
+            throws ResourceNotFoundException, MessageBusException, PersistenceException {
         AssetAdministrationShell shell = context.getPersistence().getAssetAdministrationShell(request.getId(), QueryModifier.DEFAULT);
         Reference reference = ReferenceBuilder.forAas(shell);
         if (!request.isInternal()) {