+++ core/src/main/java/de/fraunhofer/iosb/ilt/faaast/service/request/handler/submodel/AbstractInvokeOperationRequestHandler.java
@@ -21,6 +21,7 @@ import de.fraunhofer.iosb.ilt.faaast.service.model.api.Response;
 import de.fraunhofer.iosb.ilt.faaast.service.model.api.modifier.QueryModifier;
 import de.fraunhofer.iosb.ilt.faaast.service.model.api.request.submodel.InvokeOperationRequest;
 import de.fraunhofer.iosb.ilt.faaast.service.model.exception.InvalidRequestException;
+import de.fraunhofer.iosb.ilt.faaast.service.model.exception.PersistenceException;
 import de.fraunhofer.iosb.ilt.faaast.service.model.exception.ResourceNotFoundException;
 import de.fraunhofer.iosb.ilt.faaast.service.model.messagebus.EventMessage;
 import de.fraunhofer.iosb.ilt.faaast.service.request.handler.AbstractSubmodelInterfaceRequestHandler;
@@ -57,7 +58,7 @@ public abstract class AbstractInvokeOperationRequestHandler<T extends InvokeOper
 
 
     @Override
-    public U doProcess(T request) throws ResourceNotFoundException, InvalidRequestException {
+    public U doProcess(T request) throws ResourceNotFoundException, InvalidRequestException, PersistenceException {
         Reference reference = new ReferenceBuilder()
                 .submodel(request.getSubmodelId())
                 .idShortPath(request.getPath())