+++ core/src/main/java/de/fraunhofer/iosb/ilt/faaast/service/request/handler/submodel/InvokeOperationAsyncRequestHandler.java
@@ -25,6 +25,7 @@ import de.fraunhofer.iosb.ilt.faaast.service.model.api.operation.OperationResult
 import de.fraunhofer.iosb.ilt.faaast.service.model.api.request.submodel.InvokeOperationAsyncRequest;
 import de.fraunhofer.iosb.ilt.faaast.service.model.api.response.submodel.InvokeOperationAsyncResponse;
 import de.fraunhofer.iosb.ilt.faaast.service.model.exception.InvalidRequestException;
+import de.fraunhofer.iosb.ilt.faaast.service.model.exception.PersistenceException;
 import de.fraunhofer.iosb.ilt.faaast.service.model.exception.ResourceNotFoundException;
 import de.fraunhofer.iosb.ilt.faaast.service.model.exception.ValueMappingException;
 import de.fraunhofer.iosb.ilt.faaast.service.model.messagebus.event.access.OperationFinishEventMessage;
@@ -100,19 +101,19 @@ public class InvokeOperationAsyncRequestHandler extends AbstractInvokeOperationR
                                 ArgumentType.OUTPUT));
             }
         }
-        catch (ResourceNotFoundException | InvalidRequestException e) {
+        catch (ResourceNotFoundException | InvalidRequestException | PersistenceException e) {
             handleOperationFailure(reference, operationResult.getInoutputArguments(), operationHandle, e);
         }
 
-        context.getPersistence().save(operationHandle, operationResult);
         try {
+            context.getPersistence().save(operationHandle, operationResult);
             context.getMessageBus().publish(OperationFinishEventMessage.builder()
                     .element(reference)
                     .inoutput(ElementValueHelper.toValueMap(operationResult.getInoutputArguments()))
                     .output(ElementValueHelper.toValueMap(operationResult.getOutputArguments()))
                     .build());
         }
-        catch (ValueMappingException | MessageBusException e) {
+        catch (ValueMappingException | MessageBusException | PersistenceException e) {
             LOGGER.warn("could not publish OperationFinishedEventMessage on messagebus", e);
         }
     }
@@ -121,20 +122,21 @@ public class InvokeOperationAsyncRequestHandler extends AbstractInvokeOperationR
     private void handleOperationInvoke(Reference reference,
                                        OperationHandle operationHandle,
                                        InvokeOperationAsyncRequest request) {
-        context.getPersistence().save(
-                operationHandle,
-                new OperationResult.Builder()
-                        .inoutputArguments(request.getInoutputArguments())
-                        .executionState(ExecutionState.RUNNING)
-                        .build());
+
         try {
+            context.getPersistence().save(
+                    operationHandle,
+                    new OperationResult.Builder()
+                            .inoutputArguments(request.getInoutputArguments())
+                            .executionState(ExecutionState.RUNNING)
+                            .build());
             context.getMessageBus().publish(OperationInvokeEventMessage.builder()
                     .element(reference)
                     .input(ElementValueHelper.toValueMap(request.getInputArguments()))
                     .inoutput(ElementValueHelper.toValueMap(request.getInoutputArguments()))
                     .build());
         }
-        catch (ValueMappingException | MessageBusException e) {
+        catch (ValueMappingException | MessageBusException | PersistenceException e) {
             LOGGER.warn("could not publish OperationFinishedEventMessage on messagebus", e);
         }
     }