+++ core/src/main/java/de/fraunhofer/iosb/ilt/faaast/service/request/handler/submodel/PostSubmodelElementByPathRequestHandler.java
@@ -20,6 +20,7 @@ import de.fraunhofer.iosb.ilt.faaast.service.model.IdShortPath;
 import de.fraunhofer.iosb.ilt.faaast.service.model.api.modifier.QueryModifier;
 import de.fraunhofer.iosb.ilt.faaast.service.model.api.request.submodel.PostSubmodelElementByPathRequest;
 import de.fraunhofer.iosb.ilt.faaast.service.model.api.response.submodel.PostSubmodelElementByPathResponse;
+import de.fraunhofer.iosb.ilt.faaast.service.model.exception.PersistenceException;
 import de.fraunhofer.iosb.ilt.faaast.service.model.exception.ResourceAlreadyExistsException;
 import de.fraunhofer.iosb.ilt.faaast.service.model.exception.ResourceNotAContainerElementException;
 import de.fraunhofer.iosb.ilt.faaast.service.model.exception.ResourceNotFoundException;
@@ -54,7 +55,7 @@ public class PostSubmodelElementByPathRequestHandler extends AbstractSubmodelInt
     @Override
     public PostSubmodelElementByPathResponse doProcess(PostSubmodelElementByPathRequest request)
             throws ResourceNotFoundException, ValueMappingException, ValidationException, ResourceNotAContainerElementException, AssetConnectionException, MessageBusException,
-            ResourceAlreadyExistsException {
+            ResourceAlreadyExistsException, PersistenceException {
         ModelValidator.validate(request.getSubmodelElement(), context.getCoreConfig().getValidationOnCreate());
         IdShortPath idShortPath = IdShortPath.parse(request.getPath());
         Reference parentReference = new ReferenceBuilder()